Completed Authorization

+++ The-MEAN-Stack/meanCourse/backend/routes/posts.js
@@ -39,11 +39,15 @@ router.put('/:id', checkAuth, multer({storage: storage}).single('image'), (req, 
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    console.log(result);
-    res.status(200).json({message: 'Update successful!'});
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({message: 'Update successful!'});
+    } else {
+      res.status(401).json({message: 'Not Authorized!'});
+    }
   });
 });
 
@@ -55,7 +59,8 @@ router.post('', checkAuth, multer({storage: storage}).single('image'), (req, res
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + '/images/' + req.file.filename
+    imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId
   });
   console.log(post);
   // One cool amazing mongoose method
@@ -68,7 +73,8 @@ router.post('', checkAuth, multer({storage: storage}).single('image'), (req, res
         id: createdPost._id,
         title: createdPost.title,
         content: createdPost.content,
-        imagePath: createdPost.imagePath
+        imagePath: createdPost.imagePath,
+        creator: createdPost.creator
       }
     });
   });
@@ -133,9 +139,12 @@ router.get('', (req, res, next) => {
 
 // Dynamic path segment
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
-    console.log(result);
-    res.status(200).json({message: 'Post deleted!'});
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+    if (result.n > 0) {
+      res.status(200).json({message: 'Deletion successful!'});
+    } else {
+      res.status(401).json({message: 'Not Authorized!'});
+    }
   });
 });
